backend/requirements.txt
agno==2.0.7
agnos==1.0.2
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
base128==0.1.1
beautifulsoup4==4.13.5
bitarray==3.7.2
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
crypto==1.4.1
cryptography==46.0.2
ddgs==9.6.0
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
lxml==6.0.1
Markdown==3.9
markdown-it-py==4.0.0
mdurl==0.1.2
Naked==0.1.32
openai==1.108.0
oracledb==3.4.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
playwright==1.55.0
primp==0.15.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
selenium==4.35.0
shellescape==3.8.1
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
